# improve: performance for CAMM processing