… mapillary_tools/geotag/utils.py
… mapillary_tools/mp4/mp4_sample_parser.py
… mapillary_tools/telemetry.py
… mapillary_tools/video_data_extraction/extractors/camm_parser.py
… tests/cli/blackvue_parser.py
… tests/cli/camm_parser.py
… tests/unit/test_camm_parser.py