Optimizations in this PR:

1. extract GPS and camera make/model in one file open session
3. skip samples smaller than 16 bytes which are unlikely to be GPS samples
4. parse samples of interest only


```
❯ sz ~/Downloads/CAMM                                                        
9.0G ~/Downloads/CAMM
```

Before:

```
❯ time uv run -m tests.cli.camm_parser ~/Downloads/CAMM
uv run -m tests.cli.camm_parser ~/Downloads/CAMM  7.76s user 0.17s system 86% cpu 9.135 total
```

After:

```
❯ time uv run -m tests.cli.camm_parser ~/Downloads/CAMM
uv run -m tests.cli.camm_parser ~/Downloads/CAMM  3.23s user 0.16s system 81% cpu 4.182 total
```